worker/requirements.txt
redis
boto3
python-dotenv
